core/requirements.txt
flask
flask-restfuly
pandas
sqlalchemy
mysql-connector-python
